extract visitor from ExpressionAnalyser [CLICKHOUSE-3996]
