+++ dbms/src/Interpreters/interpretSubquery.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <Parsers/IAST.h>
+#include <Interpreters/InterpreterSelectWithUnionQuery.h>
+
+namespace DB
+{
+
+class Context;
+
+std::shared_ptr<InterpreterSelectWithUnionQuery> interpretSubquery(
+    const ASTPtr & table_expression, const Context & context, size_t subquery_depth, const Names & required_source_columns);
+
+}
